requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
Werkzeug==3.0.3